ASMS_Project_Folder/requirements.txt
apturl==0.5.2
asgiref==3.7.2
Babel==2.12.1
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.1.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
dj-database-url==2.0.0
Django==4.2.2
duplicity==0.8.21
fasteners==0.14.1
fonttools==4.41.0
future==0.18.2
gpg===1.16.0-unknown
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
imutils==0.5.4
jeepney==0.7.1
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.7.2
migrate==0.3.7
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.25.1
oauthlib==3.2.0
olefile==0.46
opencv-python==4.8.0.74
packaging==23.1
paramiko==2.9.3
pexpect==4.8.0
phonenumbers==8.13.16
Pillow==10.0.0
protobuf==3.12.4
psycopg==3.1.9
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pycairo==1.20.1
pycountry==22.3.5
pycparser==2.21
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.4
systemd-python==234
tkcalendar==1.6.1
torbrowser-launcher==0.3.3
typing_extensions==4.6.3
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
vine==5.0.0
wadllib==1.3.6
wcwidth==0.2.6
whitenoise==6.5.0
xdg==5
xkit==0.0.0
zipp==1.0.0
